Core Feature: Hitbox & Hurtbox
This commit implements a hitbox/hurtbox system. Hitboxes can be attached to any entity, and they keep track of any hurtboxes that enter their Area2D. They emit a signal when a hurtbox enters them and set a cooldown until the next trigger. This commit features this ability with the player character specifically.