Core Feature: Implements hurtbox damage/knockback
This commit adds a few more parameters to a hurtbox bhy allowing it to specify
the damage and knockback it would give to whatever it hits. This would let a
player controller better implement a damage and knockback system by shifting the
decider of the damage/knockback closer to the source of the hit (namely, the
hurtbox itself).