Core Features: bugfixes and hurtbox type
This commit adds a few features to the base game. First,
it fixes a simple bug with the sprite not flipping when moving left.
It fixes the bug where the origin for knockback was not being
calculated correctly. It also adds types to hurtboxes, so
that on top of blacklisting specific hurtboxes, a hurtbox
can ignore entire types of hitboxes (so that, for
example, an enemy hitbox can ignore other enemy
hurtboxes.